requirements.txt
42cc-pystyle==0.0.15
coverage==3.7.1
Django==1.6.7
django-coverage==1.2.4
flake8==2.4.1
mccabe==0.3.1
pep8==1.5.7
Pillow==2.9.0
pyflakes==0.8.1
selenium==2.47.3
South==1.0.2
wheel==0.24.0
